fix minimap fit during node drags
