+++ packages/extension/tests/webview/graph/rendering/surface/owned2d/frame/performance/model.test.ts
@@ -9,8 +9,7 @@ const frame = (
   presentationTimestampMs: number,
   renderMs = 3,
   simulationMs = 2,
-  secondaryRefreshMs?: number,
-) => ({ presentationTimestampMs, renderMs, secondaryRefreshMs, simulationMs });
+) => ({ presentationTimestampMs, renderMs, simulationMs });
 
 let nextSubmissionId = 1;
 
@@ -55,22 +54,6 @@ describe('owned graph FPS monitor', () => {
       frameTimeMs: 6.5,
       renderedFps: 50,
       sampleCount: 2,
-    });
-  });
-
-  it('reports secondary refresh work separately from total frame cost', () => {
-    const monitor = createOwnedGraphPerformanceMonitor();
-
-    recordSuccessfulFrame(monitor, frame(0, 3, 2, 0.4));
-    recordSuccessfulFrame(monitor, frame(20, 5, 3, 0.8));
-
-    expect(monitor.sample()).toEqual({
-      status: 'active',
-      frameTimeMs: 6.5,
-      renderedFps: 50,
-      sampleCount: 2,
-      secondaryRefreshMs: expect.closeTo(0.6),
-      secondaryRefreshSampleCount: 2,
     });
   });
