+++ packages/extension/tests/webview/graph/rendering/surface/owned2d/frame/runtime/fixture.ts
@@ -96,7 +96,6 @@ export function runtimeFixture(renderer: OwnedWebGpuRenderer): {
     positionVersionRef: { current: 0 },
     propsRef: { current: props },
     rendererOperationalRef: { current: true },
-    secondaryRefreshMsRef: { current: undefined },
     requestFrameRef: { current: vi.fn() },
     simulationClockRef: { current: createGraphLayoutFixedTimestepClock() },
     markPerformanceIdle: vi.fn(),
